reorganize the extension structure
Previously, all the functions were in the main
duckdb_rdkit_extension.cpp file. As the program grows, it gets to be too
messy.

First pass at better file organization.
