+++ CMakeLists.txt
@@ -16,8 +16,11 @@ project(${TARGET_NAME})
 include_directories(src/include)
 
 set(EXTENSION_SOURCES 
+    src/cast.cpp
     src/duckdb_rdkit_extension.cpp
-    src/types.cpp)
+    src/mol_formats.cpp
+    src/types.cpp
+)
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES})
 build_loadable_extension(${TARGET_NAME} " " ${EXTENSION_SOURCES})
