+++ src/cast.cpp
@@ -1,0 +1,65 @@
+#include "common.hpp"
+#include "duckdb/common/operator/cast_operators.hpp"
+#include "duckdb/common/types/string_type.hpp"
+#include "duckdb/common/types/vector.hpp"
+#include "duckdb/function/cast/default_casts.hpp"
+#include "mol_formats.hpp"
+#include "types.hpp"
+#include <GraphMol/FileParsers/FileParsers.h>
+#include <GraphMol/GraphMol.h>
+#include <GraphMol/MolPickler.h>
+#include <GraphMol/SmilesParse/SmartsWrite.h>
+#include <GraphMol/SmilesParse/SmilesParse.h>
+#include <GraphMol/SmilesParse/SmilesWrite.h>
+#include <duckdb/parser/parsed_data/create_scalar_function_info.hpp>
+
+namespace duckdb_rdkit {
+
+// This enables the user to insert into a Mol column by just writing the SMILES
+// Duckdb will try to convert the string to a rdkit mol
+// This is consistent with the RDKit Postgres cartridge behavior
+void VarcharToMol(Vector &source, Vector &result, idx_t count) {
+  UnaryExecutor::Execute<string_t, string_t>(
+      source, result, count, [&](string_t smiles) {
+        auto mol = rdkit_mol_from_smiles(smiles.GetString());
+        auto pickled_mol = rdkit_mol_to_binary_mol(*mol);
+        return StringVector::AddString(result, pickled_mol);
+      });
+}
+
+bool VarcharToMolCast(Vector &source, Vector &result, idx_t count,
+                      CastParameters &parameters) {
+  VarcharToMol(source, result, count);
+  return true;
+}
+
+void MolToVarchar(Vector &source, Vector &result, idx_t count) {
+  UnaryExecutor::Execute<string_t, string_t>(
+      source, result, count, [&](string_t bmol) {
+        auto mol = rdkit_binary_mol_to_mol(bmol.GetString());
+        auto smiles = rdkit_mol_to_smiles(mol);
+        return StringVector::AddString(result, smiles);
+      });
+}
+
+bool MolToVarcharCast(Vector &source, Vector &result, idx_t count,
+                      CastParameters &parameters) {
+  MolToVarchar(source, result, count);
+  return true;
+}
+
+void RegisterCasts(DatabaseInstance &instance) {
+  ExtensionUtil::RegisterCastFunction(instance, LogicalType::VARCHAR,
+                                      ::duckdb_rdkit::Mol(),
+                                      BoundCastInfo(VarcharToMolCast), 1);
+
+  ExtensionUtil::RegisterCastFunction(instance, duckdb_rdkit::Mol(),
+                                      LogicalType::VARCHAR,
+                                      BoundCastInfo(MolToVarcharCast), 1);
+
+  ExtensionUtil::RegisterCastFunction(instance, LogicalType::VARCHAR,
+                                      duckdb_rdkit::Mol(),
+                                      BoundCastInfo(VarcharToMolCast), 1);
+}
+
+} // namespace duckdb_rdkit
