+++ src/include/types.hpp
@@ -4,5 +4,6 @@
 namespace duckdb_rdkit {
 
 LogicalType Mol();
+void RegisterTypes(DatabaseInstance &instance);
 
 } // namespace duckdb_rdkit
