+++ src/types.cpp
@@ -1,6 +1,7 @@
 // This file defines the new types in a way for duckdb to recognize
 
-#include "types.hpp"
+#include "common.hpp"
+#include "duckdb/main/database.hpp"
 
 namespace duckdb_rdkit {
 
@@ -12,6 +13,11 @@ LogicalType Mol() {
   auto blob_type = LogicalType(LogicalTypeId::BLOB);
   blob_type.SetAlias("Mol");
   return blob_type;
+}
+
+void RegisterTypes(DatabaseInstance &instance) {
+  // Register Mol type
+  ExtensionUtil::RegisterType(instance, "Mol", duckdb_rdkit::Mol());
 }
 
 } // namespace duckdb_rdkit
